% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{Convert Two Currencies Exchange Rate}
\arguments{
\item{from}{String with the currency to change from. The format should be "XXX", for example "USD"}

\item{to}{String with the currency to change to. The format should be "XXX", for example "CAD"}
}
\value{
It will return a character vector of length two. The first element is the exchange rate. The second element is the date.
If there is an error, it will return error and get the error message.

the column will be the specific date currency exchange rate and the row will be the aim currency for that exchange rate.
}
\description{
Convert Two Currencies Exchange Rate
}
\details{
This function need the user input the currency to change from and the currency to change to.
}
\examples{
\dontrun{
convert_currency("CAD", "USD")
}
}
\references{
\url{https://exchangerate.host/#/}
}
