% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_time_series.R
\name{currency_time_series}
\alias{currency_time_series}
\title{Currency Time Series}
\arguments{
\item{start}{String with the start date. The format should be "yyyy-mm-dd", for example "2021-12-31"}

\item{end}{String with the end date. The format should be "yyyy-mm-dd", for example "2021-12-31"}

\item{base}{String with the code of the base currency. Defaults to CAD, the canadian dollars.}
}
\value{
It will return a dataframe of the currency exchange rate of the base currency in the time period.
If there is an error, it will return error and get the error message.

the column will be the specific date currency exchange rate and the row will be the aim currency for that exchange rate.
}
\description{
Get currency exchange rate of the base currency in the time series period input by the user.
}
\details{
This function need the user input the start date and the end date and the base currency
for getting the dataframe of the currency exchange rate during the time period.
}
\examples{
currency_time_series("2021-12-12", "2021-12-15", "USD")
currency_time_series("2021-01-01", "2022-01-01")

}
\references{
\url{https://exchangerate.host/#/}
}
