% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_trend_plot.R
\name{history_trend_plot}
\alias{history_trend_plot}
\title{History Trend Plot

Get the plot of the exchange rate between 2 currencies in the time period.}
\arguments{
\item{start}{String with the start date. The format should be "yyyy-mm-dd", for example "2021-12-31"}

\item{end}{String with the end date. The format should be "yyyy-mm-dd", for example "2021-12-31"}

\item{base}{String with the code of the base currency. Defaults to CAD, the canadian dollars.

@param aim String with the code of the aim currency. Defaults to CAD, the canadian dollars.}
}
\value{
It will return a plot of the currency exchange rate between the base currency and the aim currency in the time period.
If there is an error, it will return error and get the error message.

the plot x-axis will be the date and y-axis will be the currency exchange rate.
}
\description{
History Trend Plot

Get the plot of the exchange rate between 2 currencies in the time period.
}
\details{
This function need the user input the start date, end date, base currency and the aim currency
for getting the plot of the currency exchange rate between the 2 currencies during the time period.
}
\examples{
\dontrun{
history_trend_plot("2021-12-12", "2021-12-15", "USD","CAD")
history_trend_plot("2021-01-01", "2022-01-01","USD")
}
}
\references{
\url{https://exchangerate.host/#/}
}
