% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_rates.R
\name{latest_rates}
\alias{latest_rates}
\title{Latest Rates}
\arguments{
\item{base}{Enter the three-letter currency code of your preferred base currency (e.g., base='USD'),
the default is base='EUR'.}

\item{symbols}{Enter a list of comma-separated currency codes to limit output currencies (e.g., symbols=c('USD','EUR','CZK')),
the default is all of the 170+ currencies.}

\item{amount}{The amount to be converted (e.g., amount=1200), the default is amount=1.}
}
\value{
It will return a dataframe with the columns \code{Date}, \code{Base}, \code{Rates}, and \code{Amount}.
If there is an error, it will return error and get the error message.
}
\description{
Get the latest foreign exchange reference rates of a list of currencies
on a specific base at a specific amount on daily basis from the Exchange rates API.
}
\details{
This function need the user input the base currency, symbols the list of
currencies, and the amount for getting the reference rates.
}
\examples{
\dontrun{
latest_rates('CNH','USD,EUR,CAD',5000)
latest_rates(symbols='CAD,USD,CNH')
}
}
\references{
\url{https://exchangerate.host/#/}
}
